Move publish sidebar logic to core/editor store
This logic is used by the PostPublishPanel, which lives in editor.
We don't want editor to depend on anything from edit-post.

An alternative approach could be to keep the logic in edit-post
and use the the PrePublish and PostPublish extensions.
In my view, those are meant for allowing plugins to inject their
own panels into the PublishPanel. Dismissing the panel is a core
action and using the extensions to circumvent the editor/edit-post
separation would be a convoluted way to add support for dismissing the panel.