Preventive measure
In theory, isPublishSidebarEnabled shouldn't be ever undefined
in the preferences object, so this shouldn't be necessary.

In practice, we use a localStorage mechanism to persist
parts of the state in between sessions (like whether the user
has already seen the tips or not). For the core/editor store
we persist the preferences key, which contains an object.
This is the problem: the localStorage mechanism doesn't check
whether the cached preferences object and the state preferences object
have the same shape, it just overwrites the current state,
so new added keys are deleted.

While this bug is not fixed and deployed,
we want this preventive measure in place.